server/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0
faster-whisper>=1.0.0
python-Levenshtein>=0.25.0
python-multipart>=0.0.9
